Split ReservationInfoModal container & component
Refs #543.
